algumas melhorias com automação

+++ jogo-oito/src/main/java/util/WindowSize.java
@@ -11,6 +11,6 @@ public enum WindowSize {
     }
 
     public Integer getMeasure() {
-        return (Integer) value;
+        return value;
     }
 }
